requirements.txt
numpy==2.0.0
numba==0.58.1
scikit-learn>=0.24.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
vosk>=0.3.42
arabic-reshaper>=2.1.3
python-bidi>=0.4.2
scipy>=1.7.0
librosa==0.10.1
sounddevice>=0.4.4
PyQt6>=6.4.0
pyqtgraph>=0.13.1